LayoutTests/fast/canvas/webgl/array-constructor-expected.txt

Verifies that when constructing WebGLArray with nan values, it results in a 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 10 is 10
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 10 is 10
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS 0 is 0
PASS successfullyParsed is true

TEST COMPLETE
